Created AbstractRect super-class, merged functionality.
